Retry composing bundle if it's modified during setup (#2890)
### Motivation

If someone switches branches while we're setting up the composed bundle and the switch results in gemfile/lockfile changes, we often fail to boot because the resulting bundle is invalid.

### Implementation

I think we can reduce the number of failures by checking if the contents of the gemfile or lockfile have been changed after we finished running bundle install. If that's the case, we can re-build the composed bundle from scratch and retry a single time, which should be able to fix common issues.

### Automated Tests

Added a test that verifies the new behaviour.