Limit supported client schemes to file and git (#2889)
### Motivation

With the new AI developments for editors, the LSP receives a lot of unnecessary requests coming from "ghost" schemes. They are usually some sort of resource URI related to the chat or something happening in the background, but it's not actually relevant for the user.

In addition to leading to extra work in the server that is completely unnecessary, it also often confuses users because some of these resources don't send a `textDocument/didOpen` notification before starting to request for other features, which leads to the non existing document error and then people think something is broken, when in reality it's not.

### Implementation

Let's limit the schemes we handle to `file` and `git`.

### Automated Tests

Updated existing tests.