Make `enabled_feature?` return true when all flags are enabled (#2900)
### Motivation

When fixing the mistake on the feature flags object composition, I also noticed that the global state wouldn't return `true` if the user had their configuration set as

```json
{
  "rubyLsp.featureFlags": {
    "all": true
  }
}
```

which is not correct. If they enabled all flags, then checking with `enabled_feature?` should return `true`.

### Implementation

Started returning `true` if `all` is enabled.

### Automated Tests

Added a test.