Extract Ruby fallback cancellable flow (#2911)
### Motivation

This PR makes the cancellable logic used in Chruby activation reusable, so that we can use it for two purposes

1. The fallback for when no `.ruby-version` is found (already in place)
2. The fallback for when the version specified in `.ruby-version` is not installed

The cancellable flow does a lot of little things, like waiting a few seconds, allowing the user to cancel and then define fallbacks for the system or for the Ruby LSP and implementing it twice would result in a lot of duplication.

### Implementation

This PR just extracts the cancellable flow into a function we can customize.