Fix minitest spec code lens load path (#2706)
* Set load path based on test/spec directory

Some codebases (particularly those using minitest/spec) have tests in
the `spec` directory instead of the `test` directory. This commit adds
support for detecting these paths and adjusting the load path
accordingly.

I opted to not include any directory by default. If the path contains
`test` or `spec`, the appropriate directory is added to the load path.
If the path doesn't contain either, then it seems unlikely that the
`test` directory should be added to the load path.

* Use File.fnmatch? to select test directory

Co-authored-by: Vinicius Stock <vinistock@users.noreply.github.com>

---------

Co-authored-by: Vinicius Stock <vinistock@users.noreply.github.com>