Rollout launcher mode to 5% of users (#2914)
### Motivation

We've been using the new launcher mode for the server for a few weeks now, with zero issues encountered so far.

Let's roll it out to 5% of users. A larger group of users will help surface any other errors in our telemetry.

### Implementation

Instead of having a setting dedicated to the launcher, I just made it a feature flag, which is what we should do going forward.