Replace IndexablePath with URI (#2916)
### Motivation

Second step for #1908

This PR gets rid of `IndexablePath` and standardizes on using URI. This is necessary because `IndexablePath` is coupled with file paths on disk, which is incompatible with the URI scheme for unsaved files (`untitled:Untitled-1`) - a scenario where the files don't exist on disk.

### Implementation

The important changes are just in the URI extension file. Everything else is essentially just renaming things from `IndexablePath` to URI.

**Note**: this PR takes a slightly different approach than my previous attempt. I previously tried to create a separate URI class for the indexer, fully disconnected from the original `URI::Generic` class. That turned out to be a terrible idea that lead to an enormous amount of duplication, so the approach here is just to extend `URI::Generic` a bit more to handle require paths.

### Automated Tests

Added a couple of new tests for the URI extension.