Store URIs in entries instead of file paths (#2926)
### Motivation

Another step towards #1908

Entries need to store URIs instead of file paths. Otherwise, we cannot represent an unsaved file (since there's no file path associated to it, only a URI).

### Implementation

This PR changes entries to store URIs, but still maintains the assumption that every entry is backed by a file path on disk. This allows us to minimize the number of changes for this PR.

Once this goes in, then we can start looking into indexing unsaved files, which use the `untitled` scheme instead of `file` and then we will need to handle the possibility of the file path being `nil`.