Allow opting into specific feature flags (#2931)
### Motivation

In the implementation of feature flags, I forgot to check if the user opted into a specific flag. Currently, the code doesn't check for that, which means you can only opt into all of them.

### Implementation

We just need to check if the user opted into that specific flag.

### Automated Tests

Added a test.