Avoid excluding transitive dependencies of default dependencies (#2937)
### Motivation

I noticed that we were incorrectly excluding dependencies if they were transitive to both the `development` and the default groups. For example:

```ruby
gem "irb"

group :development do
  gem "debug"
end
```

In this case, both `debug` and `irb` depend on `reline`. Since `irb` is a part of the default group, we should not exclude `reline`. The bug is that we weren't taking the transitive dependencies of the default group into account and so when we saw that `debug` was in the development group, we excluded it.

### Implementation

Started concatenating the transitive dependencies that cannot be excluded into the list.

### Automated Tests

Added a test that fails before the fix.