Print debugger output to the Ruby LSP output channel as well as the debug console (#2957)
Because the debug console is only available when the debugger is actually
spawned and attached to, users often miss crucial debugging messages about
the reason the debugger exits unexpectedly.

This commit prints all output to the Ruby LSP output channel as well as the
debug console. So when the debugger exits unexpectedly, users can still see
the output in the Ruby LSP output channel.