Add a new command to help users migrate launch.json configurations
Ruby LSP now provides better logging and error messages when the debugger
failed to start. And we plan to keep improving the debugging experience
in the future.

If users are currently using vscode-rdbg and want to try Ruby LSP, we can
help them migrate their launch.json configurations to the format used by
Ruby LSP.