Infer types for method calls on immediately instantiated objects (#3007)
### Motivation

Closes #2994

This PR adds inference for methods invoked directly on an object instantiation. It's fairly easy to add, so I think it's worth getting it done.

### Implementation

When the receiver of a method call is another method call using `new`, we infer the type of the constant receiver and then return the attached version of it. For example:

```ruby
# In this case, `Foo::<Class:Foo>` is the receiver of `new`, but new will return an
# object. Thus the right type is the attached class, which is `Foo`
Foo.new.something
```

The only exception is if the method `new` was overridden, which is valid in Ruby. We cannot guarantee that the override actually returns a new object of the class, so it's better to not infer anything in those cases.

### Automated Tests

Added tests.