Use local Ruby LSP executable on node tests (#3016)
### Motivation

Our node tests use the released version of the `ruby-lsp` gem, which is not a good choice. Yesterday, only after the release of v0.23, tests started failing and indicating a bug for Windows.

We need to run the node tests against the development version of the gem, so that we test the development version of the composed bundle logic.

### Implementation

Instead of installing `ruby-lsp` from rubygems, I started packaging the gem and installing that local version.