Index documents on modification (#2941)
### Motivation

Closes #1908

This PR starts indexing files upon modification in addition to on saves.

### Implementation

The idea is to use the declaration listener together with all of the other listeners, so that we collect declaration changes within the same AST traversal. This allows us to catch any declaration modifications made before saving a file.

The biggest concern with this change is performance, but I think the next PR in this stack is doing a decent job at minimizing reindexes.

### Automated Tests

Added tests.