Go back to returning `nil` for cancelled requests (#3024)
### Motivation

This is a temporary revert to fix #3019 while we understand the correct response flow for retried cancelled requests.

This PR reverts the behaviour introduced in #2939, which didn't cause problems for NeoVim.

### Implementation

I didn't perform a full revert of every single change, since we do want to ensure that we're following the spec. I just stopped returning the error and went back to returning a `nil` response until we figure out what's going on.

### Automated Tests

Changed the test so that it asserts the current behaviour. We'll switch it back later.