Ensure requests are not assuming the presence of file paths for entries (#3030)
### Motivation

One more bug fix to include in the next release 😅. In #2941, we started indexing unsaved files. However, we didn't fix the requests that were assuming the presence of a file path in entries with `T.must`.

We need to handle the possibility of there not being a file path, so that we can account for entries in unsaved files.

### Implementation

Fixed all of the occurrences I could find where we assumed the file path to be there.

### Automated Tests

Added tests that reproduce each one of the bugs.