Fix visibility stack handling for private_class_method (#3032)
### Motivation

Found this bug while testing out things in Core. The handling of the visibility stack for `private_class_method` was incorrect. It was popping the stack once for each argument. With 3 or more arguments, that would fully exhaust the stack and cause `NoMethodError` for invocations on a `nil`.

### Implementation

I reworked the implementation because it was a bit difficult to follow. We now:

1. Only push to the visibility stack if there's a method definition as an argument. Popping happens when we leave the call node
2. Building a single list of names and process them all in one go, without nested loops

### Automated Tests

Added a test that fails on main.