Avoid reporting the same Bundler install error twice (#3033)
### Motivation

If bundle install fails and one of the gems ends up not being present, then `Bundler.setup` is guaranteed to fail with `Bundler::GemNotFound`, but that's expected. The root cause of the problem is the fact that installation failed.

Let's not set `setup_error` in that particular case to avoid reporting the same issue twice to telemetry.

### Implementation

Started not setting `setup_error` if there was an `install_error` and the error during setup matches the expected class.