Always apply workspace URI configuration to indexer (#3034)
### Motivation

Fixes #2365

We regressed on the fix to use the workspace URI given by the editor as the entry point for indexing. The changes to receive indexing configurations from the editor started returning early if no configurations are passed, but we still need to apply the workspace URI even if the user didn't specify any other configs.

### Implementation

Started applying the workspace URI to the indexing config every time.

### Automated Tests

Added a test that reproduces the bug.