Only load install error instance after requiring Bundler (#3043)
### Motivation

Closes #3038, closes #3042

I made a mistake in #3033. We cannot invoke `Marshal.load` before requiring Bundler, because then the error classes that we're trying to load will not be defined.

### Implementation

Moved the `install_error` handling after requiring Bundler and ensured that the local variable is available within scope for instantiating the server at the end of the file.