Return early if magic source link URI has no path (#3056)
### Motivation

This error was caught by our telemetry. If someone is manually typing a magic source link or if there's some error in Tapioca's generation, then the URI's path might be `nil`, which then fails when we pass it to `CGI.unescape` with a `no implicit conversion of nil to String`.

### Implementation

Started checking if there's a path before trying to unescape it.

### Automated Tests

Added a test that fails in main.