Fix singleton indexing under compact namespaces (#3062)
### Motivation

Fixes one of the bugs reported in #3050

We were assigning incorrect names to singleton classes defined inside compact namespaces. The issue is that we were simply taking the last entry in the stack, but when using compact namespaces that entry may include multiple names.

### Implementation

Started ensuring we use only the last part of the name to define the singleton.

I also noticed that we were pushing and popping from the name stack when entering a singleton method defined with `def self.foo`. That's actually unnecessary. The name stack is only used for constants and inside a method definition we'll only find instance variables (which use the owner stack rather than the name stack).

I removed the pushing and popping in that case.

### Automated Tests

Added a test.