Use local executables when working on the Ruby LSP (#3065)
### Motivation

I was trying to make some more improvements to our launcher and I realized we were missing something we should have probably done about 2 years ago.

When working on the LSP itself, we don't use the local version of the launcher. We use the local version of the server, but only after the bundle has been composed. Composing the bundle uses whatever latest version is globally installed.

This is detrimental to us catching bugs in the launch process before releasing it to users. We need to change that.

### Implementation

To ensure that the local version of the executable is always chosen when working on the LSP, I started mutating the `PATH` to include the `exe` directory.

That way, whenever `ruby-lsp` is invoked, the first thing that will be found is the local executable.

### Automated Tests

Added a test to make sure the `PATH` is being mutated.