Upgrade RuboCop related gems and Tapioca (#3069)
### Motivation

I wanted to upgrade RuboCop to the latest version so that we can verify #3067 is working. After upgrading, I noticed that Tapioca was broken because of our version of Bundler, which is fixed in a recent release - so I upgraded that too.

Then I updated our gem RBIs and fixed any issues. I split commit by commit to make it easier to review.