Add experimental/gotoRelevantFile lsp method
Define and implement the experimental/gotoRelevantFile lsp method. This
request takes in a textDcoument param, and retusn an array of locations
that are related to the input file path.

Currently, the related file paths are
- test files, if the input file is a source file
- source files, if the input file is a test file

Basically it goes between source file <> test file.

This is implemented as a experimental method until something more
official comes along.