Check if lockfile is valid before composing bundle (#3077)
### Motivation

After #3066 moves forward, we can start getting smarter about our checks of when it's valid to restart. For example, we don't need to go all the way to composing the bundle, if we already know the lockfile's syntax is invalid.

### Implementation

Started parsing the lockfile before composing the bundle, so that we can more quickly reject a restart request if there's an issue.

### Automated Tests

Added a test.