Avoid failing if file is quickly created and deleted (#3105)
### Motivation

I noticed that we were crashing if a file was quickly created and deleted. Before we got to processing the created notification, the file would already no longer be there and trying to read from disk would raise.

This can happen when running tests that automatically created/delete files or if switching branches quickly.

### Implementation

Just started rescuing `Errno::ENOENT`. If the file is no longer there, it probably means we just haven't processed its deleted notification.

### Automated Tests

Added a test that fails before the fix.