Allow snippets for Cursor editor (#3109)
### Motivation

Fixes the behaviour described in https://github.com/Shopify/ruby-lsp/discussions/3094#discussioncomment-11997299.

While the LSP spec doesn't standardize snippet support, we use a custom middleware to allow moving the user's cursor around when applying on type formatting. We were checking to apply this only in VS Code, but not Cursor.

### Implementation

Added Cursor to the mix.

### Automated Tests

Added a test.