Search for class variables completion candidates in attached namespace (#3100)
### Motivation

Class variables can be referenced both from singleton and attached contexts. To save memory, we decided to always associate class variables with only the attached version of namespaces, but we didn't account for this in either resolving class variables or finding completion candidates.

### Implementation

We need to always search for class variables within the attached namespace's ancestor chain.

### Automated Tests

Added tests that reproduce the bugs.