Fix local variable assignment highlighting with no spaces (#3131)
### Motivation

Closes #2892

The way we were matching local variables was causing assignments with no spaces after the equal sign to be incorrectly matched (see #2892 for examples).

### Implementation

Used a positive lookahead to check if the equal sign is there without matching it as part of the capture, which makes the rest of the line be processed properly.

### Automated Tests

Added a new test covering this scenario.