Handle invalid URIs in magic source links (#3132)
### Motivation

Another error caught by telemetry. If the magic source link is written manually, the user may enter something that's not a valid URI and that crashes the document link request.

### Implementation

Started handling the URI parsing errors so that we simply don't do anything if it's not a valid URI.

### Automated Tests

Added a test that reproduces the behaviour.