Prevent workspace dependencies from failing when gems aren't installed (#3138)
### Motivation

Caught by our telemetry. With the new launcher mode, we still run the language server even if gems are not installed, which makes the workspace dependencies request fail because `Bundler.definition` will raise.

### Implementation

If gems are not installed, we can't provide the dependencies view, so I just rescue the errror.

### Automated Tests

Added a couple of new tests.